requirements.txt
opencv-python>=4.8
numpy>=1.24
scikit-image>=0.21
scikit-learn>=1.3
scipy>=1.11
matplotlib>=3.7
pandas>=2.0
pyyaml>=6.0
jupyter>=1.0
ipykernel>=6.0
